Bird Collision Start
Changed bird movement to have bird move with finger specifically, instead of trying to move bird by change in X,Y from wherever finger touches. We can change this later if we choose to.

Added the code sample given to us that should check for a collision between birds, and added the code to loop over all birds and compare with placed one. Currently this returns true when no collision is obvious, need to check this.